Renamed permissions.vue to roster.vue
