+++ tests/test_course_admin/test_course_permissions/test_student_roster.ts
@@ -1,4 +1,4 @@
-import Permissions from '@/components/course_admin/roster/permissions.vue';
+import Permissions from '@/components/course_admin/roster/roster.vue';
 import StudentRoster from '@/components/course_admin/roster/student_roster.vue';
 import { config, mount, Wrapper } from '@vue/test-utils';
 import { Course, Semester, User } from 'ag-client-typescript';
